// WFHMonitor/ClientApp/src/types/api.ts
export type UnknownRecord = Record<string, unknown>;
// ...
export interface SelectOption {
  value: string;
  text: string;
  selected?: boolean;
  disabled?: boolean;
}